requirements.txt
# Core
torch
pillow
duckdb
openai
transformers
xformers
beautifulsoup4
huggingface-hub
langchain==0.0.212
chromadb==0.3.26
llama-cpp-python==0.1.65

# Get lit, get streamlit!
streamlit

# Utilities
python-dotenv
requests
colorama
